fix(types/jsx): move JSX DOM types back to `@vue/runtime-dom` (#7979)
