+++ packages/vue/types/jsx-register.d.ts
@@ -2,5 +2,3 @@
 // imports the global JSX namespace registration for compat.
 // TODO: remove in 3.4
 import '../jsx'
-
-export * from '../jsx-runtime/dom'
